Rename package to `instructlab.sdg` from `instructlab_sdg`
Part of issue #2

This changes to use Python namespace packages where `instructlab` is
the namespace used for our libraries.
