+++ .gitignore
@@ -1,5 +1,5 @@
 # Auto generated
-src/instructlab_sdg/_version.py
+src/instructlab/sdg/_version.py
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
