+++ src/instructlab/sdg/__init__.py
@@ -1,0 +1,1 @@
+from instructlab.sdg.generate_data import generate_data
