--- src/instructlab_sdg/__init__.py
@@ -1,1 +1,0 @@
-from instructlab_sdg.generate_data import generate_data
